Convert all distances into mm

+++ src/position.h
@@ -1,13 +1,15 @@
 #ifndef __POSITION_H
 #define __POSITION_H
 
+#include <stdint.h>
+
 void *position_thread(void *arg);
 void update_theta(int angle);
 void update_coordinate(int distance);
 void get_obst_position(int r, int theta, int16_t *x_obst, int16_t *y_obst);
 
 typedef struct coordinate_t{
-    int x, y, theta;
+    int16_t x, y, theta;
     pthread_mutex_t coordinate_lock;
 } coordinate_t;
